rename actions package
