+++ internal/processing/admin/account_test.go
@@ -53,7 +53,7 @@ func (suite *AccountTestSuite) TestAccountActionSuspend() {
 
 	// Wait for action to finish.
 	if !testrig.WaitFor(func() bool {
-		return suite.state.Actions.TotalRunning() == 0
+		return suite.state.AdminActions.TotalRunning() == 0
 	}) {
 		suite.FailNow("timed out waiting for admin action(s) to finish")
 	}
